created card component and displaying first promo
