+++ EcoScan/src/app/landing-page/pages/landing-pages/landing-pages.component.html
@@ -1,6 +1,8 @@
 <div [class.active]="isOpen" class="blur"></div>
 <div class="landing">
   <app-header-nav (boolChange)="onReceivedFromHeader($event)"></app-header-nav>
+  
+  <app-card/>
 
   <app-footer
     [scanButton]="true"
