+++ EcoScan/src/app/shared-module/components/feature/card/card.component.scss
@@ -1,0 +1,15 @@
+.promo-card {
+  display: flex;
+  flex-direction: column;
+  border: 2px black solid;
+  width: fit-content;
+  margin: 1rem;
+  padding: 1rem;
+  border-radius: 10px;
+  box-shadow: 4px 4px 4px grey;
+}
+
+p {
+  width: fit-content;
+  margin: 0.2rem;
+}
