+++ EcoScan/src/app/shared-module/shared-module.module.ts
@@ -1,23 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CardlistComponent } from './sliding-cards/cardlist/cardlist.component';
-
 import { SharedModuleRouting } from './shared-module-routing.module';
 import { FooterComponent } from './components/feature/footer/footer.component';
-import { CardComponent } from './sliding-cards/card/card.component';
+import { CardComponent } from './components/feature/card/card.component';
 
 @NgModule({
   declarations: [
-    CardComponent,
-    CardlistComponent,
     FooterComponent,
+    CardComponent,
   ],
   imports: [
     CommonModule,
     SharedModuleRouting,
   ],
   exports:[
-    FooterComponent
+    FooterComponent,
+    CardComponent
   ]
 })
+
 export class SharedModule { }
