+++ EcoScan/src/app/shared-module/shared/card.service.ts
@@ -12,7 +12,7 @@ export class CardService {
   // je crée mon service, smart et dumb component ✅
   // je type les données qui vont affluer ✅
   // je m'occupe de la logique métier qu'il va y avoir dans
-  // mon service, importer httpClient et faire méthodes
+  // mon service, importer httpClient et faire méthodes ✅
   // j'envoie mon service dans mon smart component (injection)
   // je m'occupe de l'affichage coté html
   // je crée un input coté enfant (dumb component)
@@ -26,7 +26,6 @@ export class CardService {
   getPromos$(): Observable<Promo[]> { 
    return this.http.get<Promo[]>(`${this.BASE_URL}/promo`).pipe(
     map((promo => promo)),
-    tap(promo => console.log(promo))
    )
 
   }
