--- EcoScan/src/app/shared-module/sliding-cards/card/card.component.html
@@ -1,1 +1,0 @@
-<p>card works!</p>
