--- EcoScan/src/app/shared-module/sliding-cards/card/card.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-card',
-  templateUrl: './card.component.html',
-  styleUrl: './card.component.scss'
-})
-export class CardComponent {
-
-}
