--- EcoScan/src/app/shared-module/sliding-cards/cardlist/cardlist.component.html
@@ -1,3 +1,0 @@
-<ng-container *ngIf="cardList$ | async as cards">
-  <div *ngFor="let card of cards">{{card.title}}</div>
-</ng-container>
